refactor: refactor back scroll render controller
